+++ providers.tf
@@ -1,9 +1,9 @@
 terraform {
-  required_version = ">= 1.6.6, < 1.8.0"
+  required_version = ">= 1.6.6, < 1.9.0"
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 5.22"
+      version = "~> 5.50"
       configuration_aliases = [
         aws.east,
         aws.main
@@ -16,7 +16,7 @@ terraform {
 
     random = {
       source  = "hashicorp/random"
-      version = "~> 3.5"
+      version = "~> 3.6"
     }
 
     archive = {